+++ testing/upload_validation.ts
@@ -2908,6 +2908,20 @@ ${endpointKey ? 'endpointKey is set' : `requiresEndpointUrl is ${requiresEndpoin
         path: ['networkDomains'],
       },
     )
+    .refine(
+      (data: any) => {
+        if (data.networkDomains?.length || !data.mcpServers?.length) {
+          return true;
+        }
+        return false;
+      },
+      {
+        message:
+          'This pack uses MCP servers but did not declare a network domain. ' +
+          "Specify the domain that your pack makes http requests to using `networkDomains: ['example.com']`",
+        path: ['networkDomains'],
+      },
+    )
     .superRefine((untypedData, context) => {
       const data = untypedData as PackVersionMetadata;
 
@@ -2977,6 +2991,29 @@ ${endpointKey ? 'endpointKey is set' : `requiresEndpointUrl is ${requiresEndpoin
               return;
             }
           }
+        }
+      }
+    })
+    .superRefine((untypedData, context) => {
+      const data = untypedData as PackVersionMetadata;
+      if (!data.networkDomains || !data.mcpServers) {
+        return;
+      }
+
+      for (const [i, server] of data.mcpServers.entries()) {
+        const usedNetworkDomain = URLParse(server.endpointUrl).hostname;
+        if (!usedNetworkDomain) {
+          continue;
+        }
+        if (
+          !data.networkDomains.some(domain => domain === usedNetworkDomain || usedNetworkDomain.endsWith('.' + domain))
+        ) {
+          context.addIssue({
+            code: 'custom',
+            path: [`mcpServers[${i}].endpointUrl`],
+            message: `Domain ${usedNetworkDomain} is used in MCP server ${server.name} but not declared in the pack's "networkDomains".`,
+          });
+          return;
         }
       }
     })